BE: Bor sung trường cho Payslip - chưa khắc phục lỗi PDF Payslip

+++ backend/src/main/java/com/manaplastic/backend/DTO/payroll/PayrollDTO.java
@@ -16,7 +16,7 @@ public class PayrollDTO {
 
     private String payPeriod;
     private BigDecimal baseSalary;
-    private Double actualWorkDays;
+    private BigDecimal actualWorkDays;
 
     private BigDecimal totalOvertimePay;
     private BigDecimal totalAllowance;
